Only delete video if it exists during tv show sync

+++ components/MovieDb.php
@@ -1312,7 +1312,9 @@ class MovieDb
 								break;
 							case 'deleted':
 								$video = ShowVideo::findOne($item->original_value->id);
-								$video->delete();
+
+								if ($video !== null)
+									$video->delete();
 
 								break;
 							default:
